♻️ refactor: 운영진, 일반 타입 분리
